refactor(desktop): split library import flow hooks
