+++ apps/desktop/src/features/library-workspace/import/model/types.ts
@@ -1,3 +1,5 @@
+export type ImportStep = 'folder' | 'assets' | 'completed';
+
 export type ImportSummary = {
   importedCount: number;
   reusedCount: number;
